requirements.txt
cassandra-driver
python-snappy
lz4
